requirements.txt
fastapi
uvicorn[standard]
python-dotenv
httpx
openai
python-multipart
Pillow
requests


